+++ Sources/MapGL/Extensions/JS+Extensions.swift
@@ -8,6 +8,20 @@ protocol IJSValue {
 	func jsValue() -> String
 }
 
+protocol IJSOptions: IJSValue {
+	func jsKeyValue() -> [String: IJSValue]
+}
+
+extension IJSOptions {
+	func jsValue() -> String {
+		let values = self.jsKeyValue()
+			.sorted { $0.key < $1.key }
+			.map { "\($0.key):\($0.value.jsValue())" }
+			.joined(separator: ",")
+		return "{\(values)}"
+	}
+}
+
 extension Optional: IJSValue where Wrapped: IJSValue {
 	func jsValue() -> String { self?.jsValue() ?? JS.undefined }
 }
@@ -33,6 +47,15 @@ extension Bool: IJSValue {
 	func jsValue() -> String { self ? "true" : "false" }
 }
 
+extension UIImage: IJSValue {
+	func jsValue() -> String {
+		guard let imageData = self.pngData() else { return JS.undefined }
+		let imageString = imageData.base64EncodedString()
+		let markerImage = "\"data:image/png;base64,\(imageString)\""
+		return markerImage
+	}
+}
+
 extension Int: IJSValue {
 	func jsValue() -> String { "\(self)" }
 }
@@ -53,6 +76,20 @@ extension PolylineStyle: IJSValue {
 
 extension Optional where Wrapped == PolylineStyle {
 	func jsValue() -> String {
-		return self?.jsValue() ?? "\(JS.undefined),\(JS.undefined),\(JS.undefined)"
+		self?.jsValue() ?? "\(JS.undefined),\(JS.undefined),\(JS.undefined)"
 	}
+}
+
+extension UIEdgeInsets: IJSValue {
+	func jsValue() -> String {
+		"[\(self.top),\(self.right),\(self.bottom),\(self.left)]"
+	}
+}
+
+extension CGSize: IJSValue {
+	func jsValue() -> String { "[\(self.width),\(self.height)]" }
+}
+
+extension CGPoint: IJSValue {
+	func jsValue() -> String { "[\(self.x),\(self.y)]" }
 }